+++ lib/core/exception.dart
@@ -1,9 +1,0 @@
-class AuthException {
-  final String msg;
-  final String? code;
-
-  const AuthException(this.msg, [this.code]);
-
-  @override
-  String toString() => "$AuthException(code: $code, msg: $msg)";
-}